modify:webhook-not found

+++ app/models.py
@@ -1,11 +1,11 @@
-from sqlalchemy import Column, Integer, String, DateTime, func, Text
+from sqlalchemy import BigInteger, Column, Integer, String, DateTime, func, Text
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql.schema import ForeignKey
 from app.database import Base
 
 
 class BaseMixin:
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     created_at = Column(DateTime, nullable=False, default=func.utc_timestamp())
     updated_at = Column(DateTime, nullable=False, default=func.utc_timestamp(), onupdate=func.utc_timestamp())
